requirements.txt
pymongo
pydantic_settings
gTTS 
cloudinary
python-dotenv
google-genai